This is a combination of 5 commits. This is the 1st commit message:
Add security multiparts for MIME (RFC 1847)

This is the commit message #2:

Implement Display for MimeMultipartType

to assign a Header the multipart type as String.

This is the commit message #3:

Add method to obtain non-Content-Type headers

to be able to add the boundary parameter to the Content-Type header
and replace all headers.

This is the commit message #4:

Stop adding a Content-Type header to add boundary

Instead, add the boundary to the Content-Type header if it exists
and instead of check that it is a multipart message by the
`message_type`, assume it's multipart if it has children.
This way there is no need to set `message_type`, it's not assumed
to be `multipart/mixed` if it has a Content-Type and the Content-Type
can have other parameters.

This is the commit message #5:

Add test with `multipart/encrypted`

Parse a message, convert to string, parse again.

Closes #50.